build: move to using cmake super build
Instead of having two separate cmake invocations, this patch adds
a single CMakeLists.txt which adds two external projects, one for
the native build, and one for the os cross compiled build. This
should be a lot cleaner, but probably still requires some work to
integrate properly with the toolchain bootstrap.