libc: move all core system libraries into libc
Because libpthread and libc were separate libraries, it was very
difficult for them to coordinate. In particular, things like tls
had to be handled in libc, which meant libc has to know about each
thread's control block. This patch merges libpthread into libc, as
well as libdl, librt, and libm.

Since everyone still wants to link against these libraries, empty
shared and static libraries are provided. This change was inspired
by glibc, which apparently plans to do this.