libedit+librepl: implement basic fuzzy matching for suggestions
Instead of doing a "true" fuzzy match, we instead use a greedy
algorithm which accepts strings where every character the user types
is matched with one character in the suggestion.

So, the string "sh" matches
    - "sh_repl.cpp"
    - "sh_parser.cpp"
    - "show_help.cpp"
but not a typoed string like "hs_repl.cpp".

However, this algorithm is comparably very efficent, and appears to
be what vscode does.