libedit+librepl+sh: refactor the suggestion offset concept
Instead of storing a suggestion offset (byte relative based on the
entire document's contents), this instead stores the text index
from which the suggestion should be inserted. The range between
this index and the cursor corresponds to the current token which
is used to generate the suggestion. This is much more efficent as
it removes the need to convert the entire document into a string.

This patch does remove some functionality from the shell, since
it uses the syntax highlighting info to determine the current
token. This currently doesn't include the token type, but this
can be fixed later. As such, it can no longer tell whether or
not a token is the first word (aka the command) or if its a token
for redirection/keyword/operator.