libtui: implement a sophisticated focus mechanism
Now, panel's must opt in to become focusable. If not, the parent panel
is focused (this recurses up the parent tree). In addition, there
is a concept of a focus proxy, which lets Panel's give their focus
to another panel. This is useful for classes which wrap text boxes, and
need their underlying text box to be focused instead of themselves.