librepl: hide the suggestions panel when accepting a suggestion
It is sometimes convenient to not close the suggestions panel, since
the user's next action will be to hit tab again. However, keeping it
open is too inconvient when trying to access a directory, especially
because the first suggestion ends up being '..'. This interacts terribly
when trying to do something like 'rm -rf dir/<enter>'.