libeventloop: enable user extensible Events
Instead of having all EventTypes in a static enum class, the Event's
type is now encoded in its name field, which is a StringView which
corresponds to the Event class's static_event_name() function.

To enforce its existenec and allow for Event's to be extended, the
APP_EVENT(name) macro is introduced. This allows for eventual automatic
event registration as well as assigning an unique integer ID to each
event at run time. The latter is a more performant implemenation for
checking where to dispatch an event.