libeventloop: enable automatic deregistration of callbacks
This works by having each callback object store an optional WeakPtr
to its event listener. If this Listener no longer exists, then the
callback is not called and gets deleted.

Note that the on<> API is extended to take an Object& as its first
argument. Alternatively, an empty GlobalListenenerTag can be provided,
to indicate the listener has global scope (it was registered in main()).

Using the WeakPtr mechanism is a bit inefficent, but works for now.
One alternative is for the listener Object itself to remember each
callback it has created, which would allow deregistration to take
place in the Object's destructor. Alternatively, the destruction of
an Object could emit an event, which would be listened to by other
Objects, allowing them to delete their listeners. Both have the same
issue in dealing with cases where their a single listener Object has
multiple callbacks. In both cases, someone must make sure the
deregistration mechanism happens only once.