libeventloop: add macro based event generation mechanism
This patch leverages the previous preprocessor work to enable
automatic generation of Event classes. For now, no default methods
other then getters/setters will be generated, but the macro based
approach allows for this to be added later.

The main macro is APP_EVENT(Namespace, EventName, Base, base_fields, fields, methods)
which generates an event named EventName in the namespace Namespace which inherits
from Base. The fields are lists of tuples tuples of the from (type,name), while methods
is a list of methods which are defined on the event.

Typical usage should be like this:
APP_EVENT(Test, TestEvent, App::Event,
    (),
    (
        (String, test)
    ),
    (
        (StringView test_view() const { return test().view(); })
    )
)

This expands too:

namespace Test {
class TestEvent : public App::Event {
    APP_EVENT_HEADER(Test, TestEvent)

public:
    explicit TestEvent(String test)
        : App::Event(static_event_name())
        , m_test(move(test))
    {}

    const String& test() const { return m_test; }
    void set_test(String test) const { m_test = move(test); }

    StringView test_view() const { return test().view(); }

private:
    String m_test;
};
}

100% of the difficultly of implementing this macro was dealing with
outputting the correct number of commas in the constructor. Preventing
trailing commas required a lot of debuggin and inventing the LIIM_LIST
mechanism.