terminal: split the TerminalWidget into 2 classes
The core Terminal functionality, which for now consists primarily
of selecting/copying/pasting text, as well as forwarding ptmx input
data to the TerminalInputParser, is now in the BaseTerminalWidget class.

Although its name includes Widget, it doesn't inherit from it at all.
Instead, it contains some pure virtual functions which convert from
Widget coordinates into terminal cells. The TerminalWidget class now
inherits from the Base class and overrides these methods.

This factoring will enable a TerminalPanel subclass, which functions
renders the TTY state directly to another terminal (like tmux or
GNU screen). In addition, the TerminalPanel will be embeddable within
other TUI applications, like the edit.

This is the first implementation of a new design pattern designed to
allow code sharing between libtui and libapp. This first pass lets
the base class register its handlers by the subclass providing an
implementation of this_widget(), which returns an *this as an
App::Object&. An alternitive approach could use composition, where
the pure virtuals would not be overriden directly but instead be
members of some type of "TerminalWidgetImpl" class. It remains to
be seen how best to architect this code, but splitting core functionality
and rendering is generally useful just from a separation of concerns
standpoint, as well as enabling sophisticated applications to be run
as a GUI and in a terminal.