libedit: let the Document directly listen for panel UI events
This gets rid of one side of the notification problem. The Document's
notifications to Display's will be tackled next.

One issue faced when performing this refactoring is that the virtual
this_widget() function as well as the Object& it returns are *cannot*
be used from the Display's destructor. This currently causes a leak
if a Document were to outlive a Panel. Its unclear how to fix this
issue, but since notifications are going to be handled by libeventloop
in the future (which uses WeakPtr), this won't be an issue (hopefully).