libapp: copy core libapp classes into separate "base" classes
The components in the App::Base namespace are designed to not in
any way dictate the way in which its Widget's are rendererd. However,
everything else should be taken care of.

Currently, the App::Base::Window class lacks clarity, as in libtui,
there is only one window possible (since there is no window management).
The Window class can be made much more generic by attempting to implement
basic windowing in the terminal. But instead, the current approach
makes will assume only 1 window in libtui.

Note this patch does not integrate things into the rest of libapp or
libtui, and just commits unused code.