libedit: implement cursor undo (control+u) (like vscode)
This works by simply taking a full cursor snapshot before each input
event. If the cursor didn't move, then the snapshot is not made.

Anytime the document is modified, the entire cursor history is cleared,
since inserting/removing characters changes the meaning of the stored
cursor text indices. This behavior is identical to vscode.

Also, the history may need to be limited, but since cursor snapshots
are not very memory intensive, this is not implemented currently.