libeventloop: add static type checking when listening for events
The new macro APP_EMITS(Parent, Events...) explicitly declares
that this object can emit these events. Any caller who attempts
to listen to an unspecified event will fail to compile.

This requires the Object's type to be known at compile time. A
run time check could be performed instead using a virtual function,
but this static approach should be more helpful.

Code can explicitly opt-out of this static checking by calling
the on_unchecked version of these functions. This is mainly required
when the object's type is not correctly narrowed. For example, Edit::Display
returns this_widget() as an App::Object instead of an App::Base::Widget.
This is done because libedit should not link against libapp (since it
will be used to power some sort of App::TextBox).

This static checking helps to address the main shortcoming with this
dynamic event handler approach by making the API exposed by some class
(like Edit::Document) fully explicit. The list of events this object
will generate is now listed at the top of its definition, and the
compilier enforces this.