libeventloop: allow coroutines to run on the event loop
This initial implementation scopes coroutines to a particular
object, which owns the coroutine. It may also make sense to have
"global" coroutines, which are owned by the event loop.

The main API added is Object::start_coroutine(), which takes a
coroutine function and calls it (producing a Task<>), and then
schedules it for execution. The key API currently missing some
way to a coroutine to wait for a specific event to be dispatched
by the event loop. Doing so will allow blocking (in a coroutine
context) while still keeping the main event loop running.