libeventloop: implement Object::until_event<Ev>()
From within a coroutine, one can perform something to the effect of

  auto event = co_await object.block_until_event<Ev>(/* coroutine owner */);

This suspends the currently executing coroutine and establishes an
event listener for the specified event, which will resume said
coroutine and automatically deregister the listener.

The current implementation ties the lifetime of every coroutine to
some specific Object. As such, this information must be propagated to
the block_until_event<Ev>() call so that a destroyed coroutine is not
resumed.

To make this fact explicit, the Object::start_coroutine() function
now takes a specialzied coroutine, App::ObjectBoundCoroutine. Currently,
this must be used in combination with Object::start_coroutine(), which
sets the owner field of the ObjectBoundCoroutine. This is used to
automatically remove itself from the owning object when the coroutines
execution completes.

It may be better to accept a Task<> instead in start_coroutine, and
then construct a wrapper ObjectBoundCourtine which simply awaits
the given Task<>. Since currently, one ObjectBoundCoroutine could
not await on another, and making on is inconvient.

This patch also makes Object::start_coroutine() take a Coroutine
directly instead of a Function. This is because mixing lambdas
and coroutines is extremely error prone, as the coroutine does
not save lambda captures itself, which means they will go out
of scope before the coroutine finishes executing. This can be
solved with some sort of wrapper mechanism, but this can be
explored later.