edit+libedit: replace callback in Display::prompt() with Task<>
Now that coroutines are properly supported, these callbacks can
be rewritten using co_await and coroutines. The key idea is that
TerminalPrompt can block until it emits a particular event. This
limits the coroutine code to only code paths which want to block.
The core component logic is still explicitly event based.