edit: avoid circular references in Terminal{Prompt,Search}
There's no need to take a strong reference to the Document in event
listeners on the Document itself. This leaked memory since the owner
of the Listener held a SharedPtr<> reference to itself. Even though
the owner was destructed, its listener remained. Since the Document
became disconnected, it never recieved any events, and thus the gc
of listeners never removed the stale listeners, even though they were
bound to a dead object.

It would be nice if Object's automatically removed their listeners from
other Object's, but it is unclear how to do this.