libliim: add Utf8View
This class implements a mechanism to iterate over a UTF-8 string
to get a series of code points.

The iterator provides two mechanism for gettings code points:
    Utf8ViewIterator::opreator*() will always return a value, U+FFFD if invalid
    Utf8ViewIterator::current_code_point() returns a Maybe<uint32_t>

This class does not handle any complicated unicode things, but simply allows
for code points to be retrieved.

Additionally, the Utf8View::validate() function will return true if and only
if the data is strictly valid UTF-8.