libedit: remove granular line_rect invalidations
This relied on computing the absolute row position of a line, which
in tern meant computing every line above it's rendered contents. This
broke MultiCursor::invalidate_based_on_last_snapshot(), which temporarily
changed the MultiCursor in order to do invalidation.

One possible fix would be to cache the absolute row position, which
may be complicated to keep accurate.

The alternative is to make cursor/selection/highlighting changes not
relayout every effected line, and instead simply patch the metadata.

In the future, both should be implemented, as the current implementation
is needlessly expensive.