libedit: fix Cursor::move_perserving_selection()
The Selection is supposed to be synced with the Cursor, but this area
ignored it. The real issue is that Selection::clear() puts the Selection's
bounds at { 0, 0 }, which is not synced with the cursor.

Ideally, the Selection class should be removed, and replaced by a single
TextIndex member in cursor. This prevents the two from getting out of
sync. Code which operates on ranges should take a start and end TextIndex
explicitly, or a different synthetic class than Selection.