libedit+edit: implement Control+\ split display
Finally. Almost all of the architectural changes made to libedit were
done to support this feature. This included moving line rendering, cursors,
and display settings out of Document, as well as implementing asynchronous
notifications (events). In fact, the entire libtui was done just to support
this feature, as previously, TerminalDisplay just rendered straight into the
terminal, emitting raw escape sequences which cleared entire lines at a time.
The TerminalRenderer abstracts and bounds all rendering to a rect, so that
multiple can appear on the screen. Merging libapp and libtui made layout/focus
trivial. libtinput's TerminalInputParser made handling Control+\ take a single
conditional, where as the previous terminal input parsing code was a complete
mess.

This commit basically means pressing Control+\ adds a new display horizontally.
This shows the same document that was being editing, and the two share state.
Pressing Control+O lets you change to a different file. Pressing Control+Q exits
the current display, and the application if none are left. For now, switching
between displays must be done using the mouse, but keyboard shortcuts should/will
be added.

Note that this patch adds a "hacked" version of Document::unregister_display(),
which won't call remove_listener() on the display. The issue is that upon display
destruction, it must unregister itself from the Document, so its doesn't keep a stale
reference. However, since its running its destructor, the virtual this_widget() function
cannot be called without invoking UB (manifesting as calling __cxa_pure_virtual). This
listeners are automatically removed via scoping to WeakPtr's, the listeners only need to
be removed if the Display is going to start listening to another document. Its likely
that this can be fixed by simply removing the Display registration system, and replacing
it with more events. The listeners should also be replaced by the keyboard shortcut mechanism
added a while back.

Next up is adding the libtui backed terminal to this thing!