libedit+edit: use new key bindings mechanism
The purpose of the commit is to prepare for user editable key bindings.
This will probably manifest itself as either a simple tabular text file,
like:

```
Select All:control+a
Scroll Up:control+up
```

or some sort of JSON format, like:
```
{
    "Select All": {
        "shortcut": "control+a"
    },
    "Go To Line": {
        "shortcut": "control+g"
    }
}
```

or perhaps the JSON would be structured as modifiers+key explicitly.

Ideally most of this serialization would be handled by libeventloop,
but its hard to know what API this should be before it actually gets
used. So for now, this patch just replaces the KeyDownEvent handler
with a KeyBindings object.