edit: add control+t shortcut to open/close an embedded terminal
Unfoutunately, control+` cannot be used because the terminal sends
a NUL byte in this case (which is supposed to be control+space).
As such, control+t will be used until I think of something better.

Additionally, the editor crashes when the shell inside the terminal
exits, but this should be fixed in a later commit.