edit: add multi shortcut to focus the integrated terminal
This patch moves the split_display() logic from inside TerminalDisplay
and into a event handler setup by the parent. This makes it easier to
reference the necessary variables, and allows custom logic to be injected
by the parent. In this case, this functionality is used to let the
integrated terminal be conditionally focused by control+k control+down.