libliim+libext: refactor Ext::MappedFile to store a ByteBuffer
This should make code which parses binary file formats (like PNG),
be able to easily worked with MappedFile's or data accumulated over
the network (stored in a ByteBuffer). These formats are not well suited
to using the stream API if they require the entire file to be parsed,
usually because they have raw byte offsets in the file format.