libedit: introduce and use the CommandGroup class
A lot of commands can be conviently be expressed as a sequence of
more primitive commands or cursor movement. For example, search and
replace is equivalent to selecting every search match, and then
inserting the replacement string. When undoing this action, the fact
that a selection occured needs to be hidden. The CommandGroup ensures
this by logically starting the command before the cursor movement,
instead or after.

This additionally provides a mechanism to combine consecutive commands
together. A simple example is combining multiple insertions into one
logical command, so that the undo happens sooner. A more complex example
is language aware editing which can surround the selected text with
double quotes (effectively 2 insertions) instead of replacing the
selection with double quotes.