libedit: add "insert around" behavior for certain characters
When selecting text, it is often disirable to insert matching
characters around the selection, for instance, to make something
into a string (by pressing <double-quote>). This could be implemented
as a separate command, but this implementation uses command groups
(which are infinitely more flexible). To handle cursor movement in
between commands, the MovementCommand is added, which does nothing
but call a specified lambda which should move cursors around).