libeventloop: add Object::intercept<Event>()
This functions equivalently to Object::on<Event>(), but registers
the handler before others instead of after. This is a somewhat hackish
solution to proper event ordering, but it does seem quite workable.

I believe this is really the only way to control ordering when using
this kind of declarative event handling system. The alternative to
this approach is the explicitly state the ordering of events, which
requires all events to be known by whoever dispatches them. This is
not at all feasible with this design.