libedit: don't recompose lines when only the metadata changes
Previously, a change to syntax highlighting would require recomposing
the entire document, in order for the new syntax highlighting information
to take effect. This patch prevents this by only invalidating the metadata,
not the line's composition.

Now, every time a line is rendered, its metadata is recomputed to reflect
any metadata changes. Since this doesn't require allocation, this is a lot
lighter than a full recomposition. This could be furthered improved to
invalidate metadata more precisely, but this already is a large performance
win.

Note that the line's composition is purely for display purposes, like
word wrapping or showing inline suggestions.