libedit: use binary search in TextRangeCollection::iterator()
This functions needs to find the correct range index to start iteration
at, given a start TextIndex. The previous implementation used a linear
scan, but this is too slow for large files.

Since the ranges are sorted, we can use binary search to reduce the
time complexity from O(N) to O(log N), which is a huge different since
this function is called for every line which appears on the screen.