libedit: remove the Selection class
The Selection class is made obsolete by the newer TextRange class,
which works better for anything that wants to represent a range of
text. This refactor's main goal is to ensure that the Selection is
always valid. Previously, code had to randomly ensure that an
empty Selection would have its start and end equal to the cursor's
index.

This patch removes these fixups by constructing the selection TextRange
directly from the Cursor, so it will always be valid. A Cursor now
stores an optional "selection anchor", which when set to value indicates
the origin point of the selection. If not present, the Selection will
always be empty, but the "selection anchor" is still reported as the
cursor's index. Once set, the selection becomes the text range between
the cursor's current index and the stored "selection anchor".

This results in a simplification of any code which updated the Selection,
since it is now done automatically.