libedit: optimize Document::scroll_cursor_into_view()
When the cursor is very far from the Display's bounding region, we
can set the scroll offsets based off the cursor's position and then
adjust if needed. This stops trying to get the display position of
an index which is very far off screen, since doing so requires
composing every line between the display's scroll offset and the
cursor's index.

This makes pressing control+end on large files instant.