libedit: add smarter functions for adding/removing lines
Since lines are stored in a vector, inserting lines is going to be
O(N). However, when copy pasting large segments of text, we incur
the cost of moving over elements in the Vector a lot less if only
one event gets generated.

This makes use of the new Vector<T> methods to make inserting/removing
multiple lines more efficent. This doesn't solve the underlying issue
of using a Vector<> to store every line.