libedit: move line functions which deal with Display to RenderedLine
Since the actual line class must consult the Display for its
corresponding RenderedLine anyway, the RenderedLine should be
the one to handle calls related to rendering information.

The Display class provides a convenient wrapper function for
calls to the RenderedLine, which no longer must take 2 extra
paramaters for both the Document& and Display&. The Display
ensures the RenderedLine is up to date before calling any
of the Display functions.

This refactor makes all code which used rendered Line information
a lot less visually noisy.