libedit: make the asynchronous syntax highlighting smarter
This code should be architected in such a way that managing the
threads is much easier, but this works for now. This patch also
makes the highlighting synchronous for "small" documents, which is
most noticable for the shell, which definitely wants to highlight
things synchronously. This approach should be somewhat reasonable.