libapp: replace ModelData variant with ModelItemInfo class
Instead of calling a single virtual function multiple times, model
item info is now packed into one structure (with Maybe<T> instead
of a variant). This hopefully makes the code clearer, all through
the full model API hasn't been truely flushed out yet.

The item_info() virtual function takes in an integer mask which
controls which fields should be returned. This is because it would
be incredibly wasteful to do something like load an PNG icon for
a model which was being rendered by something like a terminal,
which won't display any icons.