libapp: convert models to use virtual ModelItem class
The goal is to replace the fragile ModelIndex concept with a virtual
ModelItem class. Internally, all models have some sort of Object class
which represents a piece of data (ex: FileSystemObject in FileSystemModel).
The model previously had to handle storage, management, and rendered
info lookup for each FileSystemObject.

This patch instead makes FileSystemObject inherit from ModelItem,
and has the Model class store the ModelItem's for its subclasses.
This means that code can hang on to ModelItem's directly instead
of using ModelIndexes, and sending granular Model update events
will become automatic. This factoring ensures that the Model cares
only about loading data and does not have to manage issues like
syncing Model updates to Views and Selections properly.

The only issue with this approach is that ModelItems must be
stored as a pointer instead directly in a Vector. However,
this is a performance issue which can be fixed later by having
each Model subclass customize the ModelItem fixed size, which
along with a custom Vector class, would ensure efficent storage.

This patch prepares for a hiarchical data by making every ModelItem
have a list of children, and making every Model have exactly one
root node. This tree structure will be used by the file manager to
provide some sort of TreeView, which could also be used in edit.