libapp: keep track each ModelItem's parent
This is done automatically using newly added constructor methods,
which also simplify client code. The parents must be tracked so
that tree like data can be worked with easily. For instance, the
file system model needs to reconstruct the full path of a ModelItem,
which can be done by following the parent chain.