libapp: convert the FileSystemModel to include the entire file tree
This makes use of the new View::set_root_item() to render the correct
directory. This patch also refactors the model item creation methods
to be in the Model class itself, so that the Model is aware of all
changes which occur. This means that Model's no longer have to manually
call did_update() when the Model is updated, instead, the Model will
generate this event each time a mutation method is called. Note that
this will generate a lot more events than the previous method, and
so should be worked on further, perhaps with some RAII which will
consider a set of mutations to be a "transaction" within the Model.