libapp: remove now useless ModelIndex concept
Now that everything goes through the ModelItem class, there is no
need to have a notion of a ModelItem. Instead, we can remember an
exact pointer to the item in question. This is actually very convenient
in cases where items are being added in the middle of a Model. When
using a index, the index would have to be adjusted to point to the
same thing. Now, things will work automatically.

The only drawback is that users may not listen to all model updates,
which it must do to realize that a pointer it stores has now become stale.