libeventloop: Add abstract App::Component class
The component class provides a uniform mechanism for "plug-ins" to
register themselves with a given object. Their intended usage would
be something like as follows:

class ScrollComponent : public App::Component {
public:
    ScrollComponent(Widget& widget) : App::Component(widget) {}

    int scroll_y() const { ... }
    void draw_scrollbar(...) { ... }

    virtual void did_attach() {
        object().intercept<ScrollEvent>(...);
    }

    ...
};

class View extends ... {
public:
    View(...) : ..., m_scroll_component(*this) {}

    ...

private:
    ...
    ScrollComponent m_scroll_component;
};

The idea with this API is that if the ScrollComponent will be automatically
attached to given View class, which will allow the component to setup
event handlers on behalf of the View. In the case of scrolling, it is expected
that render code will consult convenience methods provided by the ScrollComponent
to ensure that a scrolling behavior is achieved.

In general, this API is an attempt to begin fixing usages of inheritence for
code sharing in libapp and especially everything in the Base namespace. Components
would become even more powerful if they can be inheritted from directly, although
this causes problems because every component requires Object to first be initialized.
But since they are both super class, this is at least very error prone to achieve.
Perhaps a solution to this problem could be rely even more on Events, since components
can always listen in. Directly wrapping the component with forwarding accessors/setters
would also work, and could even possibly be automated using reflection or a templated
stateless base class which uses a member pointer.

Hopefully, at the very least, this can replace the current usage of having classes override
a this_widget() function, which is approximately the same as inheriting from a Component.