libapp: implement a basic ScrollComponent
Essentially, the ScrollComponent maintains variables like a
scroll offset and whether certain directions can be scrolled. It
overrides the MouseScrollEvent to change these offsets, and provides
a get_renderer() method which configures the Renderer to respect the
scroll offset and available area to draw (since this can draw scrollbars).

Currently, there is no easy way for a custom widget to interface with
this component (which can be done in the future by using virtual
inheritted methods), so the ScrollComponent makes use of the layout
constraint to determine the maximum size of a widget. Note that this
it actually good, since it will force many Widgets to actually set
these values correctly (which will improve the layout engine's results).

Additionally, this patch changes the abstract Component interface
to not have a detach() function. This is because this function is
effectively useless, since the lifetime of any Component will be
tied directly (and thus equivalent) to its Object. This also makes
things clearer given that this Component mechanism uses inheritance
for now (to allow easy exportation of Component methods).