libapp: properly handle granular updates in View and TreeView
For the View class, this means removing pointers to ModelItem
which no longer exist (like the hovered item and selected items).

For the TreeView class, this means syncing the layout items with
the actual model. This allows state to exist (like whether a folder
is open) even when the Model changes. The TreeView uses the Model's
granular updates to preserve this state. Interestingly, the problem
of syncing ModelItem's with layout state is quite similiar to the
problem of React syncing its render information with the actual DOM
tree. Its possible that granular updates could be avoided by assigning
some unique identifier to each ModelItem, and using that the synchronize
the two trees (like the React key attribute), but this would probably
make the TreeView class a lot more complicated.