libapp: make the Splitter sizes user adjustable
This patch also refactors the existing code to assume no hidden widgets
exist in the interal items vector. This makes the rest of the code
much similiar, even though it makes finding the proper place to insert
widgets when they are unhidden a bit morem difficult.

This implementation has a lot of problems, mostly concerning setting both
the default and minimum sizes of widgets in the spliiter.