meta: continue porting the system to i686
This patch does enough work to get the kenrel to compile, but not
link. Additionally, this the patchelf utility is now used to set
libgcc_s.so as a dependency of libc.so. This ensures every program
is which links against libc will get this library, and also lets
libc use libgcc functions, which is very nice for the i686 port.