kernel: begin refactoring usages of create_phys_addr_mapping()
Although the name of this function suggests that something is being
created, there was no corresponding free_phys_addr_mapping() function,
and so these were leaked all over the place. This is fine on x86_64,
since the kernel maps all of physical memory, but this is not possible
on i686, where the address space is only 32 bits.

Note that these methods are meant only for temporary physical address
mapping, and code which uses these as long lived blocks should be
converted to a separate mechanism.