kernel: make it compile and boot on i686
The kernel now runs until update_vga_buffer(), which calls into
vm_make_region(), which allocates memory. This reveals a great flaw
in the implementation, which currently does not implement the
critical to paging create_temp_phys_addr_mapping().

The paging code relies on this function, but upon reflection, this
function will probably need to call into the paging code in order
to actually work. This is a disaster.

I speculate that the easy solution will be just to use recursive
mappings specifically for the paging code. Although it may be
possible to forego using a recursive mapping, since the temp pages
will be located at fixed addresses (at least for the bsp). For other
processors this is not possible, since they will be dynamically
allocated. In effect, the paging code can't use temporary page mappings,
and thus must by recursive. This is disapointing, as recursive page mappings
are ugly.