kernel: help along the i686 boot process
We now get all the way to trying to read from the hard drive.
However, the kernel has been completely broken this entire time.
Basically, when waiting for an disk IRQ, it will block the current
task. However, the current task, during boot, is the same as the idle
task. The idle task will be scheduled by default, even if it is blocked
(in fact, it can't be blocked). Thus, the code doesn't actually block
at all. This could only possiblely work because qemu will perform the
interrupt immediately (most of the time, this probably explains some
rare random kernel crashes where it fails to read from disk).

For whatever reason, this is much more noticeable on i686, and now must
be fixed.