kernel: get all the way to starting userland for i686
This patch fixes IRQ handling, by taking care of the fact that %esp
and %ss and only pushed if a privledge level change occurs. In
addition, iret won't set %esp, so the code to resume a task needs
to set %esp to the correct value itself.

The get_phys_addr() function is also fixed under i686. It previously
forgot to take into account the page offset of the given virtual
address.

Now, the major problem is loading userspace binaries, since currently
only Elf64 is supported.