kernel: work on i686 userspace
This patch makes the elf loader work for both 32 bit and 64 bit
binaries.

It additionally fixes an error in the interrupt handler logic which
made the code sometimes fail to save the current stack pointer.
Unfortunately, there still seems to be a silent stack corruption
bug somewhere, which needs to be fixed before working on userspace
intergration.