kernel+loader+toolchain: successfully boot userland on i686
Note that since signals are still unsupported, we crash before displaying
the GUI. However, userspace tasks work, and so does fork/exec and the
dynamic loader.

This patch fixed the passing of arguments from the dynamic loader to
libc as well as a bug in the toolchain script which installed the
wrong version of libgcc_s.so.1 to the sysroot.