kernel: get i686 to an almost stable point
This patch properly implements signal dispatch of i686, which makes
userspace bootup successfully. This patch additionally makes acpi
support optional, meaning that minimalistic hardware/emulators can
be used without the OS crashing. This was a matter of disabling smp
when acpi is not detected, and not enabling things like the io apic
when acpi is not present.

However, the kernel crashes as soon as the heap expands more the 2 MiB.
The reason for this is that the kernel page tables are not properly
synchronized across all address spaces. On x86_64, the kernel vm
resides only in the last 2 PML4 entries. On x86 without PAE, the last
1 GiB of address space require ~300 PD entries. The solution to this
issue is either to properly keep track of the kernel page tables (difficult),
require PAE, which would allow the kernel to use only one PAE entry (requires
rewriting tedious paging code), or to just prefill the kernel page tables
(easy, but wastes ~1.0 MiB of physical memory) until the kernel vm is full.