kernel: port to the x86 web emulator: https://copy.sh/v86/
The main body of work was in porting the os to 32 bit x86, and this
patch consists of a few hacks make the emulated hardware work properly.
The first is that the returned device ID for the PS2 keyboard is different
that the device ids expected, the second byte is decimal 83 instead of hex
83, although as long as the first byte is 0xAB it should be treated as a
keyboard. The second issue is that the PCI IDE controller is mapped to
non-default IO ports by the BIOS, which v86 ignores. This is fixed by
passing a boot parameter to the kernel to ignore the IO ports reported
by PCI.

In addition to this kernel work, making a directly bootable image is
necessary since v86 gets confused if there is both a cdrom and a hard
disk image (it creates 2 IDE controllers, one for each, and they are
mapped to the same IO ports). A special grub config is provided for the
emulator, which sets the required command line flags so that the os runs
properly.